chore: move getAllProjectsQuery to graphql folder
