+++ web-registry/src/lib/queries/react-query/registry-server/graphql/getAllProjectsQuery/getAllProjectsQuery.ts
