+++ web-registry/src/pages/Marketplace/Storefront/Storefront.tsx
@@ -31,7 +31,7 @@ import { useLedger } from 'ledger';
 import { getHashUrl } from 'lib/block-explorer';
 import { getBatchQuery } from 'lib/queries/react-query/ecocredit/getBatchQuery/getBatchQuery';
 import { getProjectsQuery } from 'lib/queries/react-query/ecocredit/getProjectsQuery/getProjectsQuery';
-import { getAllProjectsQuery } from 'lib/queries/react-query/registry-server/getAllProjectsQuery/getAllProjectsQuery';
+import { getAllProjectsQuery } from 'lib/queries/react-query/registry-server/graphql/getAllProjectsQuery/getAllProjectsQuery';
 import { getMetadataQuery } from 'lib/queries/react-query/registry-server/getMetadataQuery/getMetadataQuery';
 import { getAllCreditClassesQuery } from 'lib/queries/react-query/sanity/getAllCreditClassesQuery/getAllCreditClassesQuery';
